+++ sdk/resourcemanager/redisenterprise/armredisenterprise/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Release History
 
+## 1.1.0 (2022-06-06)
+### Features Added
+
+- New const `NameEnterpriseE20`
+- New const `NameEnterpriseE50`
+- New const `NameEnterpriseFlashF700`
+- New const `NameEnterpriseE10`
+- New const `NameEnterpriseFlashF1500`
+- New const `NameEnterpriseFlashF300`
+- New const `NameEnterpriseE100`
+- New function `PossibleNameValues() []Name`
+- New struct `Capability`
+- New struct `LocationInfo`
+- New struct `RegionSKUDetail`
+- New struct `RegionSKUDetails`
+- New struct `SKUDetail`
+- New struct `SKUsClientListOptions`
+- New struct `SKUsClientListResponse`
+
+
 ## 1.0.0 (2022-05-17)
 
 The package of `github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/redisenterprise/armredisenterprise` is using our [next generation design principles](https://azure.github.io/azure-sdk/general_introduction.html) since version 1.0.0, which contains breaking changes.